third_party/WebKit/LayoutTests/http/tests/serviceworker/navigation-preload/request-headers.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Navigation Preload request headers</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>
<script>
promise_test(t => {
    var script = 'resources/request-headers-worker.js';
    var scope = 'resources/request-headers-scope.php';
    return service_worker_unregister_and_register(t, script, scope)
      .then(registration => {
          add_completion_callback(_ => registration.unregister());
          var worker = registration.installing;
          return wait_for_state(t, worker, 'activated');
        })
      .then(_ => with_iframe(scope))
      .then(frame => {
          var headers = JSON.parse(frame.contentDocument.body.textContent);
          assert_equals(
            headers['HTTP_SERVICE_WORKER_NAVIGATION_PRELOAD'],
            'hello',
            'The header of Navigation Preload request must have correct ' +
            '"Service-Worker-Navigation-Preload" header.');
          assert_equals(
            headers['HTTP_UPGRADE_INSECURE_REQUESTS'],
            '1',
            'The header of Navigation Preload request must have ' +
            '"Upgrade-Insecure-Requests" header.');
        });
  }, 'Navigation Preload request headers.');

</script>
